AbstractArrayStorage,SortedArrayStorage and ArrayStorage methods fixed
